requirements.txt
google-cloud-bigquery==3.20.1
db-dtypes==1.2.0
